+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+## [0.3.0] - 2023-06-15
 ### Added
 - `Audio::sample_rate()`
 - `DelayAndSum` sound source separation.
@@ -29,7 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2023-05-14
 INITIAL RELEASE
 
-[unreleased]: https://github.com/ModProg/ssloc/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/ModProg/ssloc/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/ModProg/ssloc/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/ModProg/ssloc/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/ModProg/ssloc/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/ModProg/ssloc/releases/tag/v0.1.0